requirements.txt
pandas
numpy
scikit-learn
tqdm
nltk
PyYAML
xgboost
matplotlib
seaborn